+++ src/Components/index.js
@@ -3,5 +3,5 @@ export {Comment} from "./Comments/Comments";
 export {FollowerList} from "./FollowerList/FollowerList";
 export {PostCard} from "./PostCard/PostCard";
 export {PostBox} from "./PostBox/PostBox";
-export {ProfileCard} from "./Profile/ProfileCard";
-export {ProfileModal} from "./Profile/ProfileModal"
+export {ProfileCard} from "./Profile/Card-Profile";
+export {ProfileModal} from "./Profile/Edit-Profile-Modal";
